separate timestamp tests into directory
